requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.26.0 